refactor: backend 구조 개선 및 Kysely 데이터베이스 연동 추가
- 확장 모듈들을 개별 디렉터리로 분리 (cors, kysely, open-api, scalar)
- Kysely 데이터베이스 연동 기능 추가
- 패키지 의존성 업데이트
- TypeScript 설정 및 Wrangler 설정 개선
