+++ backend/src/extensions/cors/cors.extension.ts
@@ -1,6 +1,6 @@
 import { OpenAPIHono } from '@hono/zod-openapi';
 import { cors } from 'hono/cors';
-import { Bindings } from '../types';
+import { Bindings } from '../../types';
 
 export function corsExtension(app: OpenAPIHono<Bindings>): void {
   app.use('*', cors());
