+++ backend/src/extensions/kysely/kysely.database.ts
@@ -1,0 +1,1 @@
+export interface Database {}
