+++ backend/src/types.ts
@@ -1,4 +1,11 @@
-export type Env = {};
+import { Kysely } from 'kysely';
+import { Database } from './extensions/kysely/kysely.database';
+
+export type Env = {
+  DB: D1Database;
+
+  db: Kysely<Database>;
+};
 
 export type Bindings = {
   Bindings: Env;
